Removed dependencies

+++ lib/matrix/matrix.hpp
@@ -12,7 +12,7 @@ namespace pix::adt
      * @tparam N_LIN Number of lines
      * @tparam N_COL Number of columns
     */
-    template <typename type_t, size_t N_LIN, size_t N_COL>
+    template <typename type_t, unsigned long N_LIN, unsigned long N_COL>
     class matrix
     {
     private:
@@ -45,7 +45,7 @@ namespace pix::adt
          * @tparam _N_COL Number of columns
          * @param mtx Matrix
         */
-        template <typename _type_t, size_t _N_LIN, size_t _N_COL>
+        template <typename _type_t, unsigned long _N_LIN, unsigned long _N_COL>
         matrix(matrix<_type_t, _N_LIN, _N_COL>&);
 
         /**
@@ -54,19 +54,19 @@ namespace pix::adt
          * @return Reference to line vector
          * @note Index is congruent to number of lines
         */
-        vector<type_t, N_COL>& operator [] (size_t);
+        vector<type_t, N_COL>& operator [] (unsigned long);
 
         /**
          * @brief Get number of lines
          * @return Number of lines
         */
-        constexpr const size_t n_lin(void) const;
+        constexpr const unsigned long n_lin(void) const;
 
         /**
          * @brief Get number of columns
          * @return Number of columns
         */
-        constexpr const size_t n_col(void) const;
+        constexpr const unsigned long n_col(void) const;
 
         /**
          * @brief Assignment operator
@@ -75,7 +75,7 @@ namespace pix::adt
          * @tparam _N_COL Number of columns
          * @param mtx Matrix
         */
-        template <typename _type_t, size_t _N_LIN, size_t _N_COL>
+        template <typename _type_t, unsigned long _N_LIN, unsigned long _N_COL>
         void operator = (matrix<_type_t, _N_LIN, _N_COL>&);
 
         /**
@@ -86,7 +86,7 @@ namespace pix::adt
          * @param mtx Matrix
          * @return true if matrices are equal
         */
-        template <typename _type_t, size_t _N_LIN, size_t _N_COL>
+        template <typename _type_t, unsigned long _N_LIN, unsigned long _N_COL>
         const bool operator == (matrix<_type_t, _N_LIN, _N_COL>&) const;
 
         /**
@@ -104,7 +104,7 @@ namespace pix::adt
          * @param mtx Matrix
          * @return true if matrices are different
         */
-        template <typename _type_t, size_t _N_LIN, size_t _N_COL>
+        template <typename _type_t, unsigned long _N_LIN, unsigned long _N_COL>
         const bool operator != (matrix<_type_t, _N_LIN, _N_COL>&) const;
 
         /**
@@ -157,7 +157,7 @@ namespace pix::adt
          * @param mtx Matrix
          * @return Multiplication
         */
-        template <typename _type_t, size_t _N_COL>
+        template <typename _type_t, unsigned long _N_COL>
         matrix<type_t, N_LIN, _N_COL> operator * (matrix<_type_t, N_COL, _N_COL>&) const;
     };
 }
